% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bf_func}
\alias{bf_func}
\title{Internal function used to calculate arbitrary bias factors.}
\usage{
bf_func(rr1, rr2)
}
\arguments{
\item{rr1}{A number. A risk ratio that is a component of a bias factors}

\item{rr2}{A number. The other risk ratio that is a component of a bias
factors}
}
\value{
Returns the value of the expression. Used for calculating bias
factors.
}
\description{
Internal function used to calculate arbitrary bias factors.
}
\keyword{internal}
