% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EValue.R
\name{bias_plot}
\alias{bias_plot}
\title{Plot bias factor as function of confounding relative risks}
\usage{
bias_plot(RR, xmax)
}
\arguments{
\item{RR}{The relative risk}

\item{xmax}{Upper limit of x-axis.}
}
\description{
Plots the bias factor required to explain away a provided relative risk.
}
\examples{
# recreate the plot in VanderWeele and Ding (2017)
bias_plot(RR=3.9, xmax=20)
}
\keyword{e-value}
