% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deg_func}
\alias{deg_func}
\title{Internal function used to fit roots of a polynomial made up of the product of
bias factors.}
\usage{
deg_func(x, y, n, d)
}
\arguments{
\item{x}{A number. The variable to solve for.}

\item{y}{A number. The observed risk ratio.}

\item{n}{A number. Degree of polynomial in the numerator.}

\item{d}{A number. Degree of polynomial in the denominator.}
}
\value{
Returns the value of the expression. Used for root solving. At the
function's roots, will return 0.
}
\description{
Internal function used to fit roots of a polynomial made up of the product of
bias factors.
}
\keyword{internal}
