% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{g}
\alias{g}
\title{define transformation in a way that is monotonic over the effective range of B (>1)
to avoid ggplot errors
helper function for confounded_meta}
\usage{
g(x)
}
\description{
define transformation in a way that is monotonic over the effective range of B (>1)
to avoid ggplot errors
helper function for confounded_meta
}
