% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EValue.R
\docType{data}
\name{lead}
\alias{lead}
\title{An example dataset}
\format{
An object of class \code{function} of length 1.
}
\usage{
lead
}
\description{
An example dataset from Hsu and Small (Biometrics, 2013).
}
\keyword{datasets}
