% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{Phat_causal}
\alias{Phat_causal}
\title{\subsection{Phat after shifting by bias factor B and using calibrated estimates}{

.dat needs to have a column called "calib"
helper function for confounded_meta
}}
\usage{
Phat_causal(
  .q,
  .B,
  .calib,
  .tail,
  .give.CI = TRUE,
  .R = 2000,
  .dat = NA,
  .calib.name = NA
)
}
\description{
\subsection{Phat after shifting by bias factor B and using calibrated estimates}{

.dat needs to have a column called "calib"
helper function for confounded_meta
}
}
