% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{sens_plot}
\alias{sens_plot}
\title{Plots for sensitivity analyses}
\usage{
sens_plot(
  method = "calibrated",
  type,
  q,
  CI.level = 0.95,
  tail = NA,
  give.CI = TRUE,
  Bmin,
  Bmax,
  breaks.x1 = NA,
  breaks.x2 = NA,
  muB,
  sigB,
  yr,
  vyr = NA,
  t2,
  vt2 = NA,
  R = 1000,
  dat = NA,
  yi.name = NA,
  vi.name = NA
)
}
\arguments{
\item{type}{\code{dist} for distribution plot; \code{line} for line plot (see Details)}

\item{q}{True effect size that is the threshold for "scientific significance"}

\item{CI.level}{Pointwise confidence level as a proportion}

\item{Bmin}{Lower limit of lower X-axis on the log scale (only needed for line plot)}

\item{Bmax}{Upper limit of lower X-axis on the log scale (only needed for line plot)}

\item{breaks.x1}{Breaks for lower X-axis (bias factor) on RR scale (optional for line plot; not used for distribution plot)}

\item{breaks.x2}{Breaks for upper X-axis (confounding strength) on RR scale (optional for line plot; not used for distribution plot)}

\item{muB}{Single mean bias factor on log scale (only needed for distribution plot)}

\item{sigB}{Standard deviation of log bias factor across studies (length 1)}

\item{yr}{Pooled point estimate (on log scale) from confounded meta-analysis}

\item{vyr}{Estimated variance of pooled point estimate from confounded meta-analysis}

\item{t2}{Estimated heterogeneity (tau^2) from confounded meta-analysis}

\item{vt2}{Estimated variance of tau^2 from confounded meta-analysis}
}
\description{
Produces line plots (\code{type="line"}) showing the bias factor on the relative risk (RR) scale vs. the proportion
of studies with true RRs above \code{q} (or below it for an apparently preventive relative risk).
The plot secondarily includes a X-axis scaled based on the minimum strength of confounding
to produce the given bias factor. The shaded region represents a 95\% pointwise confidence band.
Alternatively, produces distribution plots (\code{type="dist"}) for a specific bias factor showing the observed and
true distributions of RRs with a red line marking exp(\code{q}).
}
\details{
Arguments \code{vyr} and \code{vt2} can be left \code{NA}, in which case no confidence
band will appear on the line plot.
}
\examples{

##### Example 1: Calibrated Line Plots #####

# simulated dataset with exponentially distributed 
#  population effects
# we will use the calibrated method to avoid normality assumption
data(toyMeta)

# without confidence band
sens_plot( method = "calibrated",
           type="line",
           q=log(.9),
           tail = "below",
           Bmin=log(1),
           Bmax=log(4),
           dat = toyMeta,
           yi.name = "est",
           vi.name = "var",
           give.CI = FALSE )


# # with confidence band and a different threshold, q
# # commented out because takes a while too run
# sens_plot( method = "calibrated",
#            type="line",
#            q=log(1),
#            tail = "below",
#            Bmin=log(1),
#            Bmax=log(4),
#            dat = toyMeta,
#            yi.name = "est",
#            vi.name = "var",
#            give.CI = TRUE,
#            R = 300 ) # should be higher in practice


##### Example 2: Calibrated and Parametric Line Plots #####

# example dataset
d = metafor::escalc(measure="RR",
                    ai=tpos,
                    bi=tneg,
                    ci=cpos,
                    di=cneg,
                    data=metafor::dat.bcg)

# without confidence band
sens_plot( method = "calibrated",
           type="line",
           tail = "below",
           q=log(1.1),
           Bmin=log(1),
           Bmax=log(4),
           dat = d,
           yi.name = "yi",
           vi.name = "vi",
           give.CI = FALSE )

# # with confidence band
# # commented out because  it takes a while
# # this example gives bootstrap warnings because of its small sample size
# sens_plot( method = "calibrated",
#            type="line",
#            q=log(1.1),
#            Bmin=log(1),
#            Bmax=log(4),
#            R = 500,  # should be higher in practice (e.g., 1000)
#            dat = d,
#            yi.name = "yi",
#            vi.name = "vi",
#            give.CI = TRUE )


# now with heterogeneous bias across studies (via sigB) and with confidence band
sens_plot( method = "parametric",
           type="line",
           q=log(1.1),
           yr=log(1.3),
           vyr = .05,
           vt2 = .001,
           t2=0.4,
           sigB = 0.1,
           Bmin=log(1),
           Bmax=log(4) )

##### Distribution Line Plot #####

# distribution plot: apparently causative
sens_plot( type="dist",
           q=log(1.1),
           muB=log(2),
           sigB = 0.1,
           yr=log(1.3),
           t2=0.4 )

# distribution plot: apparently preventive
sens_plot( type="dist",
           q=log(0.90),
           muB=log(1.5),
           sigB = 0.1,
           yr=log(0.7),
           t2=0.2 )
}
\keyword{confounding}
\keyword{meta-analysis}
\keyword{sensitivity}
