% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\docType{data}
\name{soyMeta}
\alias{soyMeta}
\title{A meta-analysis on soy intake and breast cancer risk (Trock et al., 2006)}
\format{
An object of class \code{data.frame} with 20 rows and 3 columns.
}
\usage{
soyMeta
}
\description{
A meta-analysis of observational studies (12 case-control and six cohort or nested case-control) on the association of soy-food intake with breast cancer risk. Data are from Trock et al.'s (2006) Table 1. This dataset was used as the applied example in Mathur & VanderWeele (2020).
}
\details{
The variables are as follows:
\itemize{
\item \code{author} Last name of the study's first author.
\item \code{est} Point estimate on the log-relative risk or log-odds ratio scale.
\item \code{var} Variance of the log-relative risk or log-odds ratio.
}
}
\references{
Trock BJ, Hilakivi-Clarke L, Clark R (2006). Meta-analysis of soy intake and breast cancer risk. \emph{Journal of the National Cancer Institute}.

Mathur MB & VanderWeele TJ (2020a). Sensitivity analysis for unmeasured confounding in meta-analyses. \emph{Journal of the American Statistical Association}.
}
\keyword{datasets}
