% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{That_causal_bt}
\alias{That_causal_bt}
\title{\subsection{That and Ghat from grid search of Phat values}{for each of a vector of bias factors, calculates Phat causal and then finds the onethat's closest to threshold proportion, .r

helper function for confounded_meta
}}
\usage{
That_causal_bt(original, indices, .calib.name, .q, .r, .B.vec, .tail)
}
\description{
\subsection{That and Ghat from grid search of Phat values}{for each of a vector of bias factors, calculates Phat causal and then finds the onethat's closest to threshold proportion, .r

helper function for confounded_meta
}
}
