% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\docType{data}
\name{toyMeta}
\alias{toyMeta}
\title{An example meta-analysis}
\format{
An object of class \code{data.frame} with 50 rows and 2 columns.
}
\usage{
toyMeta
}
\description{
A simple simulated meta-analysis of 50 studies with exponentially distributed population effects.
}
\details{
The variables are as follows:
\itemize{
\item \code{est} Point estimate on the log-relative risk scale.
\item \code{var} Variance of the log-relative risk.
}
}
\keyword{datasets}
