% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{confounded_meta}
\alias{confounded_meta}
\title{Estimates and inference for sensitivity analyses}
\usage{
confounded_meta(
  method = "calibrated",
  q,
  r = NA,
  CI.level = 0.95,
  tail = NA,
  muB,
  R = 2000,
  sigB,
  yr,
  vyr = NA,
  t2,
  vt2 = NA,
  give.CI = TRUE,
  dat,
  yi.name,
  vi.name
)
}
\arguments{
\item{method}{"calibrated" or "parametric"}

\item{q}{True effect size that is the threshold for "scientific significance"}

\item{r}{For \code{Tmin} and \code{Gmin}, value to which the proportion of strong effect sizes is to be reduced}

\item{CI.level}{Confidence level as a proportion}

\item{tail}{\code{above} for the proportion of effects above \code{q}; \code{below} for
the proportion of effects below \code{q}. By default, is set to \code{above} if the pooled point estimate (\code{method == "parametric"}) or median of the calibrated estimates (\code{method == "calibrated"}) is above 1 on the relative risk scale and is set to \code{below} otherwise.}

\item{muB}{Mean bias factor on the log scale across studies. When considering bias that of homogeneous strength across studies (i.e., \code{method == "calibrated"} or \code{method = "parametric"} with \code{sigB = 0}), \code{muB} represents the log-bias factor in each study.}

\item{R}{Number  of  bootstrap  or  simulation  iterates  (depending  on  the  methods  cho-sen).   Not required if using ci.method = "parametric"and bootstrapping is not needed.}

\item{sigB}{Standard deviation of log bias factor across studies}

\item{yr}{Pooled point estimate (on log scale) from confounded meta-analysis}

\item{vyr}{Estimated variance of pooled point estimate from confounded meta-analysis}

\item{t2}{Estimated heterogeneity (tau^2) from confounded meta-analysis}

\item{vt2}{Estimated variance of tau^2 from confounded meta-analysis}

\item{give.CI}{Logical. If TRUE, bootstrap confidence intervals provided}

\item{Bmin}{Lower limit of bias factor (used for "calibrated" method only)}

\item{Bmax}{Upper limit of bias factor (used for "calibrated" method only)}

\item{calib.name}{column name in dataframe containing calibrated estimates}
}
\description{
Computes point estimates, standard errors, and confidence interval bounds
for (1) \code{prop}, the proportion of studies with true effect sizes above \code{q} (or below
\code{q} for an apparently preventive \code{yr}) as a function of the bias parameters;
(2) the minimum bias factor on the relative risk scale (\code{Tmin}) required to reduce to
less than \code{r} the proportion of studies with true effect sizes more extreme than
\code{q}; and (3) the counterpart to (2) in which bias is parameterized as the minimum
relative risk for both confounding associations (\code{Gmin}).
}
\details{
To compute all three point estimates (\code{prop, Tmin, and Gmin}) and inference, all
arguments must be non-\code{NA}. To compute only a point estimate for \code{prop},
arguments \code{r, vyr}, and \code{vt2} can be left \code{NA}. To compute only
point estimates for \code{Tmin} and \code{Gmin}, arguments \code{muB, vyr}, and \code{vt2}
can be left \code{NA}. To compute inference for all point estimates, \code{vyr} and
\code{vt2} must be supplied.
}
\examples{
d = metafor::escalc(measure="RR", ai=tpos, bi=tneg,
ci=cpos, di=cneg, data=metafor::dat.bcg)

m = metafor::rma.uni(yi= d$yi, vi=d$vi, knha=FALSE,
                     measure="RR", method="DL" ) 
yr = as.numeric(m$b)  # metafor returns on log scale
vyr = as.numeric(m$vb)
t2 = m$tau2
vt2 = m$se.tau2^2 

# obtaining all three estimators and inference
confounded_meta( q=log(0.90), r=0.20, muB=log(1.5), sigB=0.1,
                 yr=yr, vyr=vyr, t2=t2, vt2=vt2,
                 CI.level=0.95 )

# passing only arguments needed for prop point estimate
confounded_meta( q=log(0.90), muB=log(1.5),
                 yr=yr, t2=t2, CI.level=0.95 )

# passing only arguments needed for Tmin, Gmin point estimates
confounded_meta( q=log(0.90), r=0.20,
                 yr=yr, t2=t2, CI.level=0.95 )
}
\keyword{meta-analysis}
