% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{g}
\alias{g}
\title{#' Simplified Phat_causal for bootstrapping
#'
#' An internal function that calls \code{Phat_causal} in a way that plays well with the \code{boot} package. Draws a resample internally. Only returns Phat itself.
#' @param original The original dataset from which to resample (will be passed by \code{boot})
#' @param indices The indices to resample (will be passed by \code{boot})
#' @param q True effect size that is the threshold for "scientific significance"
#' @param B Single value of bias factor}
\usage{
g(x)
}
\description{
#' @param tail \code{above} for the proportion of effects above \code{q}; \code{below} for
#' the proportion of effects below \code{q}.
#' @param calib.name Column name in dataframe \code{dat} containing calibrated estimates
#' @import
#' @noRd
#' boot
Phat_causal_bt = function( original,
indices,\preformatted{                       q,
                       B,
                       tail,
                       calib.name) \{
}
}
\section{draw bootstrap sample}{

}

