% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\docType{data}
\name{toyMeta}
\alias{toyMeta}
\title{An example meta-analysis}
\format{
An object of class \code{data.frame} with 50 rows and 2 columns.
}
\usage{
toyMeta
}
\description{
An simple simulated meta-analysis with exponentially distributed population effects.
}
\keyword{datasets}
